Fixed root redirect #205
Fixes #205 - root redirect issue when Exhibitor is not installed as ROOT.war, but as a separate application, changed the absolute redirect in the default-redirect servlet filter to a relative one. Tested and the redirect works for both "/" and "/some-other-context/". The standalone uber jar doesn't redirect for "/" and that is existing behaviour(returns a 404).